Revert to 594ff86e6e394c0e28486d6e10d2e9f8b84b8d40

+++ Data/Mappers/ScopedObjects/Questions.cs
@@ -17,13 +17,6 @@ public class Questions : OLabMapper<SystemQuestions, QuestionsDto>
   {
   }
 
-  //public QuestionsPhys(
-  //  IOLabLogger Logger, 
-  //  WikiTagProvider tagProvider, 
-  //  bool _enableWikiTranslation = true) : base(Logger, tagProvider)
-  //{
-  //}
-
   public override SystemQuestions ElementsToPhys(IEnumerable<dynamic> elements, Object source = null)
   {
     var phys = GetPhys(source);
@@ -44,7 +37,7 @@ public class Questions : OLabMapper<SystemQuestions, QuestionsDto>
 
     phys.ShowAnswer = Convert.ToInt32(elements.FirstOrDefault(x => x.Name == "show_answer").Value) == 1 ? true : false;
     if (uint.TryParse(elements.FirstOrDefault(x => x.Name == "counter_id").Value, out uint id1))
-      phys.CounterId = (int)id1;
+      phys.CounterId = id1;
 
     if (int.TryParse(elements.FirstOrDefault(x => x.Name == "num_tries").Value, out int id2))
       phys.NumTries = id2;
